spartan adapter works
